pyproject.toml
[tool.poetry]
name = "smartcab-bot"
version = "0.1.0"
description = ""
authors = ["Alchemmist <anton.ingrish@gmail.com>"]
readme = "README.md"
packages = [{include = "smartcab"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.8"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.3"
redis = "^5.0.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
ruff = "^0.3.3"
pyright = "^1.1.354"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
lint.ignore = ["F401"]

